requirements.txt
streamlit>=1.28.0
st-gsheets-connection>=0.1.0
pandas>=2.0.0
pytz>=2023.3
requests>=2.31.0
gspread>=5.10.0
google-auth>=2.22.0
tomli>=2.0.1

